@@ -11,31 +11,24 @@ import { NavigationComponent } from './navigation/navigation.component';
 import { FooterComponent } from './footer/footer.component';
 import { OperationModeComponent } from './operation-mode/operation-mode.component';
 import { OperationModeService } from './services/operation-mode/operation-mode.service';
-import { HttpClientModule } from '@angular/common/http';
+import { provideHttpClient, withInterceptorsFromDi } from '@angular/common/http';
 import { FormsModule } from '@angular/forms';
 import { RobotComponent } from './devices/robot/robot.component';
 import { StorageComponent } from './devices/storage/storage.component';
 
-@NgModule({
-  declarations: [
-    AppComponent,
-    HeaderComponent,
-    MdiComponent,
-    AboutComponent,
-    HomeComponent,
-    NavigationComponent,
-    FooterComponent,
-    OperationModeComponent,
-    RobotComponent,
-    StorageComponent,
-  ],
-  imports: [
-    BrowserModule,
-    AppRoutingModule,
-    FormsModule,
-    HttpClientModule,
-  ],
-  providers: [OperationModeService],
-  bootstrap: [AppComponent]
-})
+@NgModule({ declarations: [
+        AppComponent,
+        HeaderComponent,
+        MdiComponent,
+        AboutComponent,
+        HomeComponent,
+        NavigationComponent,
+        FooterComponent,
+        OperationModeComponent,
+        RobotComponent,
+        StorageComponent,
+    ],
+    bootstrap: [AppComponent], imports: [BrowserModule,
+        AppRoutingModule,
+        FormsModule], providers: [OperationModeService, provideHttpClient(withInterceptorsFromDi())] })
 export class AppModule { }
