@@ -44,7 +44,7 @@ public class Program
 
     private static IHostBuilder CreateHostBuilder(string[] args)
     {
-        string contentDirectory = $"{DirectoryProvider.ResolveContentDirectory()}/wwwroot";
+        string contentDirectory = $"{DirectoryProvider.ResolveContentDirectory()}/wwwroot/browser";
         Log.Information("Setting the web root directory to '{ContentDirectory}'", contentDirectory);
 
         return Host.CreateDefaultBuilder(args)
